move lithic blade into correct weapon folder
