+++ internal/weapons/claymore/lithic/lithic.go
