+++ pkg/simulation/imports.go
@@ -146,6 +146,7 @@ import (
 	_ "github.com/genshinsim/gcsim/internal/weapons/claymore/debateclub"
 	_ "github.com/genshinsim/gcsim/internal/weapons/claymore/favonius"
 	_ "github.com/genshinsim/gcsim/internal/weapons/claymore/ferrousshadow"
+	_ "github.com/genshinsim/gcsim/internal/weapons/claymore/lithic"
 	_ "github.com/genshinsim/gcsim/internal/weapons/claymore/nagamasa"
 	_ "github.com/genshinsim/gcsim/internal/weapons/claymore/oldmercspal"
 	_ "github.com/genshinsim/gcsim/internal/weapons/claymore/pines"
@@ -206,7 +207,6 @@ import (
 	_ "github.com/genshinsim/gcsim/internal/weapons/sword/ironsting"
 	_ "github.com/genshinsim/gcsim/internal/weapons/sword/kagotsurubeisshin"
 	_ "github.com/genshinsim/gcsim/internal/weapons/sword/lion"
-	_ "github.com/genshinsim/gcsim/internal/weapons/sword/lithic"
 	_ "github.com/genshinsim/gcsim/internal/weapons/sword/mistsplitter"
 	_ "github.com/genshinsim/gcsim/internal/weapons/sword/primordial"
 	_ "github.com/genshinsim/gcsim/internal/weapons/sword/prototype"
